refactor: extract agendasInjector and meetingBlockManager; use stable meeting keys; centralize prune/sort
